Allow up to 2^31-1 frames in segmentations.
At several occassions dcmseg and underlying modules limited the number
of frames, e.g. by (partly) restricting this value to 16 bit.

Now in theory 2^31-1 frames are possible, though the size for
non-concatenation instances is still limited to around 4 GB imposed by
the Pixel Data element's length field for uncompressed pixel data.

Added method on segmentations and underlying Multi-frame Dimension
module that prevents running dimension checking (especially Dimension
Index Pointer attributes) which can slow down writing a lot if many
frames are written.

Also made some simplifications and added documentation.

+++ dcmseg/include/dcmtk/dcmseg/segment.h
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright (C) 2015-2021, Open Connections GmbH
+ *  Copyright (C) 2015-2022, Open Connections GmbH
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  This software and supporting documentation are maintained by
@@ -97,7 +97,7 @@ public:
     /** Get Segment Number
      *  @return The Segment number
      */
-    virtual unsigned int getSegmentNumber();
+    virtual Uint16 getSegmentNumber();
 
     /** Get Segment Label
      *  @param  value Reference to variable in which the value should be stored
